requirements.txt
fastapi~=0.110.0
boto3==1.34.65
botocore==1.34.65
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
jmespath==1.0.1
moviepy==1.0.3
numpy==1.26.4
opencv-python==4.8.1.78
pillow==10.2.0
proglog==0.1.10
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
tqdm==4.66.2
urllib3==2.2.1
typing-extensions == 4.8.0
rich==12.1.0
httpx==0.22.0
uvicorn~=0.28.1